+++ src/collection.ts
@@ -1,5 +1,6 @@
 import { Connection } from "./connection";
 import { ArrayCursor } from "./cursor";
+import { isArangoError } from "./error";
 
 export enum CollectionType {
   DOCUMENT_COLLECTION = 2,
@@ -30,7 +31,8 @@ export interface ArangoCollection {
   name: string;
 }
 
-const COLLECTION_NOT_FOUND = 1203;
+export const DOCUMENT_NOT_FOUND = 1202;
+export const COLLECTION_NOT_FOUND = 1203;
 export abstract class BaseCollection implements ArangoCollection {
   isArangoCollection: true = true;
   name: string;
@@ -49,7 +51,9 @@ export abstract class BaseCollection implements ArangoCollection {
     }
   }
 
-  protected abstract _documentPath(documentHandle: DocumentHandle): string;
+  protected _documentPath(documentHandle: DocumentHandle) {
+    return `/document/${this._documentHandle(documentHandle)}`;
+  }
 
   protected _documentHandle(documentHandle: DocumentHandle) {
     if (typeof documentHandle !== "string") {
@@ -109,10 +113,10 @@ export abstract class BaseCollection implements ArangoCollection {
     return this.get().then(
       () => true,
       err => {
-        if (err.errorNum !== COLLECTION_NOT_FOUND) {
-          throw err;
+        if (isArangoError(err) && err.errorNum === COLLECTION_NOT_FOUND) {
+          return false;
         }
-        return false;
+        throw err;
       }
     );
   }
@@ -198,6 +202,40 @@ export abstract class BaseCollection implements ArangoCollection {
       },
       res => res.body
     );
+  }
+
+  documentExists(documentHandle: DocumentHandle): Promise<boolean> {
+    return this._connection
+      .request(
+        {
+          method: "HEAD",
+          path: `/_api/${this._documentPath(documentHandle)}`
+        },
+        () => true
+      )
+      .catch(err => {
+        if (err.statusCode === 404) {
+          return false;
+        }
+        throw err;
+      });
+  }
+
+  document(
+    documentHandle: DocumentHandle,
+    graceful: boolean = false
+  ): Promise<any> {
+    const result = this._connection.request(
+      { path: `/_api/${this._documentPath(documentHandle)}` },
+      res => res.body
+    );
+    if (!graceful) return result;
+    return result.catch(err => {
+      if (isArangoError(err) && err.errorNum === DOCUMENT_NOT_FOUND) {
+        return null;
+      }
+      throw err;
+    });
   }
 
   replace(documentHandle: DocumentHandle, newValue: any, opts: any = {}) {
@@ -645,17 +683,6 @@ export class DocumentCollection extends BaseCollection {
     super(connection, name);
   }
 
-  protected _documentPath(documentHandle: DocumentHandle) {
-    return `/document/${this._documentHandle(documentHandle)}`;
-  }
-
-  document(documentHandle: DocumentHandle) {
-    return this._connection.request(
-      { path: `/_api/${this._documentPath(documentHandle)}` },
-      res => res.body
-    );
-  }
-
   save(data: any, opts?: DocumentSaveOptions | boolean) {
     if (typeof opts === "boolean") {
       opts = { returnNew: opts };
@@ -702,11 +729,11 @@ export class EdgeCollection extends BaseCollection {
     return `document/${this._documentHandle(documentHandle)}`;
   }
 
-  edge(documentHandle: DocumentHandle) {
-    return this._connection.request(
-      { path: `/_api/${this._documentPath(documentHandle)}` },
-      res => res.body
-    );
+  edge(
+    documentHandle: DocumentHandle,
+    graceful: boolean = false
+  ): Promise<any> {
+    return this.document(documentHandle, graceful);
   }
 
   save(data: any, opts?: DocumentSaveOptions | boolean): Promise<any>;