+++ src/test/19-graph-vertex-collections.ts
@@ -45,27 +45,39 @@ describe("GraphVertexCollection API", function() {
   describe("graphVertexCollection.vertex", () => {
     let data = { foo: "bar" };
     let meta: any;
-    beforeEach(done => {
-      collection
-        .save(data)
-        .then(result => {
-          meta = result;
-          done();
-        })
-        .catch(done);
+    beforeEach(async () => {
+      meta = await collection.save(data);
     });
-    it("returns a vertex in the collection", done => {
-      collection
-        .vertex(meta._id)
-        .then(doc => {
-          expect(doc).to.have.keys("_key", "_id", "_rev", "foo");
-          expect(doc._id).to.equal(meta._id);
-          expect(doc._key).to.equal(meta._key);
-          expect(doc._rev).to.equal(meta._rev);
-          expect(doc.foo).to.equal(data.foo);
-        })
-        .then(() => void done())
-        .catch(done);
+    it("returns a vertex in the collection", async () => {
+      const doc = await collection.vertex(meta._id);
+      expect(doc).to.have.keys("_key", "_id", "_rev", "foo");
+      expect(doc._id).to.equal(meta._id);
+      expect(doc._key).to.equal(meta._key);
+      expect(doc._rev).to.equal(meta._rev);
+      expect(doc.foo).to.equal(data.foo);
+    });
+    it("does not throw on not found when graceful", async () => {
+      const doc = await collection.vertex("does-not-exist", true);
+      expect(doc).to.equal(null);
+    });
+  });
+  describe("graphVertexCollection.document", () => {
+    let data = { foo: "bar" };
+    let meta: any;
+    beforeEach(async () => {
+      meta = await collection.save(data);
+    });
+    it("returns a vertex in the collection", async () => {
+      const doc = await collection.document(meta._id);
+      expect(doc).to.have.keys("_key", "_id", "_rev", "foo");
+      expect(doc._id).to.equal(meta._id);
+      expect(doc._key).to.equal(meta._key);
+      expect(doc._rev).to.equal(meta._rev);
+      expect(doc.foo).to.equal(data.foo);
+    });
+    it("does not throw on not found when graceful", async () => {
+      const doc = await collection.document("does-not-exist", true);
+      expect(doc).to.equal(null);
     });
   });
   describe("graphVertexCollection.save", () => {